refactor: components/analytics folder
