+++ components/analytics/instructions/instructions-upload.tsx
@@ -1,5 +1,5 @@
 import { Accordion } from "@radix-ui/react-accordion";
-import DefaultAccordionItem from "../ui-elements/accordion/DefaultAccordionItem";
+import DefaultAccordionItem from "../../ui-elements/accordion/DefaultAccordionItem";
 import InstructionsUploadMrr from "./instructions-upload-mrr";
 import InstructionsUploadSrrp from "./instructions-upload-srrp";
