+++ components/analytics/table/render-month-mrr.tsx
